+++ modules/tinymce/src/themes/silver/main/ts/api/Settings.ts
@@ -5,8 +5,9 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Option, Type, Obj, Arr } from '@ephox/katamari';
-import { SelectorFind, Body, Element } from '@ephox/sugar';
+import { Arr, Obj, Option, Type } from '@ephox/katamari';
+import { Body, Element, SelectorFind } from '@ephox/sugar';
+import DOMUtils from 'tinymce/core/api/dom/DOMUtils';
 import Editor from 'tinymce/core/api/Editor';
 import EditorManager from 'tinymce/core/api/EditorManager';
 import { AllowedFormat } from 'tinymce/core/api/fmt/StyleFormat';
@@ -32,11 +33,12 @@ const getSkinUrl = function (editor: Editor): string {
 const isReadOnly = (editor): boolean => editor.getParam('readonly', false, 'boolean');
 const isSkinDisabled = (editor: Editor) => editor.getParam('skin') === false;
 
-const getHeightSetting = (editor): number => editor.getParam('height', Math.max(editor.getElement().offsetHeight, 200));
-const getMinWidthSetting = (editor): Option<number> => Option.from(editor.settings.min_width).filter(Type.isNumber);
-const getMinHeightSetting = (editor): Option<number> => Option.from(editor.settings.min_height).filter(Type.isNumber);
-const getMaxWidthSetting = (editor): Option<number> => Option.from(editor.getParam('max_width')).filter(Type.isNumber);
-const getMaxHeightSetting = (editor): Option<number> => Option.from(editor.getParam('max_height')).filter(Type.isNumber);
+const getHeightSetting = (editor: Editor): string | number => editor.getParam('height', Math.max(editor.getElement().offsetHeight, 200));
+const getWidthSetting = (editor: Editor): string | number => editor.getParam('width', DOMUtils.DOM.getStyle(editor.getElement(), 'width'));
+const getMinWidthSetting = (editor: Editor): Option<number> => Option.from(editor.settings.min_width).filter(Type.isNumber);
+const getMinHeightSetting = (editor: Editor): Option<number> => Option.from(editor.settings.min_height).filter(Type.isNumber);
+const getMaxWidthSetting = (editor: Editor): Option<number> => Option.from(editor.getParam('max_width')).filter(Type.isNumber);
+const getMaxHeightSetting = (editor: Editor): Option<number> => Option.from(editor.getParam('max_height')).filter(Type.isNumber);
 
 const getUserStyleFormats = (editor: Editor): Option<AllowedFormat[]> => Option.from(editor.getParam('style_formats')).filter(Type.isArray);
 const isMergeStyleFormats = (editor: Editor): boolean => editor.getParam('style_formats_merge', false, 'boolean');
@@ -77,7 +79,8 @@ const isMultipleToolbars = (editor: Editor): boolean => {
 export enum ToolbarDrawer {
   default = '',
   floating = 'floating',
-  sliding = 'sliding'
+  sliding = 'sliding',
+  scrolling = 'scrolling'
 }
 
 const getToolbarDrawer = (editor: Editor): ToolbarDrawer => editor.getParam('toolbar_drawer', '', 'string') as ToolbarDrawer;
@@ -113,6 +116,7 @@ export {
   isReadOnly,
   isSkinDisabled,
   getHeightSetting,
+  getWidthSetting,
   getMinWidthSetting,
   getMinHeightSetting,
   getMaxWidthSetting,