+++ modules/tinymce/src/themes/silver/main/ts/modes/Inline.ts
@@ -7,15 +7,17 @@
 
 import { AlloyComponent, Attachment, Docking } from '@ephox/alloy';
 import { Option } from '@ephox/katamari';
-import { Css, Element, Height, Location } from '@ephox/sugar';
+import { Body, Css, Element, Height, Location, Width } from '@ephox/sugar';
 import DOMUtils from 'tinymce/core/api/dom/DOMUtils';
 import Editor from 'tinymce/core/api/Editor';
-import { getToolbarDrawer, getUiContainer, isStickyToolbar, ToolbarDrawer, useFixedContainer } from '../api/Settings';
+import { getMaxWidthSetting, getToolbarDrawer, getUiContainer, isStickyToolbar, ToolbarDrawer, useFixedContainer } from '../api/Settings';
 import { UiFactoryBackstage } from '../backstage/Backstage';
 import { setupReadonlyModeSwitch } from '../ReadOnly';
 import { ModeRenderInfo, RenderArgs, RenderUiComponents, RenderUiConfig } from '../Render';
 import OuterContainer from '../ui/general/OuterContainer';
 import { identifyMenus } from '../ui/menus/menubar/Integration';
+import * as EditorSize from '../ui/sizing/EditorSize';
+import Utils from '../ui/sizing/Utils';
 import { inline as loadInlineSkin } from './../ui/skin/Loader';
 import { setToolbar } from './Toolbars';
 
@@ -25,6 +27,7 @@ const render = (editor: Editor, uiComponents: RenderUiComponents, rawUiConfig: R
   const useFixedToolbarContainer = useFixedContainer(editor);
   const isSticky = isStickyToolbar(editor);
   const targetElm = Element.fromDom(args.targetNode);
+  const editorMaxWidthOpt = getMaxWidthSetting(editor).or(EditorSize.getWidth(editor));
 
   const splitSetting = getToolbarDrawer(editor);
   const split = splitSetting === ToolbarDrawer.sliding || splitSetting === ToolbarDrawer.floating;
@@ -47,6 +50,14 @@ const render = (editor: Editor, uiComponents: RenderUiComponents, rawUiConfig: R
       top: Math.round(top) + 'px',
       left: Math.round(location.left()) + 'px'
     });
+
+    // Update the max width of the inline toolbar
+    const maxWidth = editorMaxWidthOpt.getOrThunk(() => {
+      // No max width, so use the body width, minus the left pos as the maximum
+      const bodyMargin = Utils.parseToInt(Css.get(Body.body(), 'margin-left')).getOr(0);
+      return Width.get(Body.body()) - location.left() + bodyMargin;
+    });
+    Css.set(floatContainer.element(), 'max-width', maxWidth + 'px');
   };
 
   const updateChromeUi = (resetDocking: boolean = false) => {