+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/SilverInlineEditorWidthTest.ts
@@ -1,10 +1,13 @@
-import { ApproxStructure, Assertions, Chain, GeneralSteps, Logger, Pipeline, UiFinder } from '@ephox/agar';
+import { ApproxStructure, Assertions, Chain, GeneralSteps, Logger, Pipeline, Step, UiFinder } from '@ephox/agar';
 import { UnitTest } from '@ephox/bedrock';
+import { Arr, Fun } from '@ephox/katamari';
 import { TinyApis, TinyLoader } from '@ephox/mcagar';
-import { Css, Element } from '@ephox/sugar';
+import { Body, Css, Element, Scroll } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
+import { ToolbarDrawer } from 'tinymce/themes/silver/api/Settings';
 import Theme from 'tinymce/themes/silver/Theme';
+import { sOpenMore } from '../../module/MenuUtils';
 
 UnitTest.asynctest('Inline Editor (Silver) width test', (success, failure) => {
   Theme();
@@ -15,18 +18,18 @@ UnitTest.asynctest('Inline Editor (Silver) width test', (success, failure) => {
       ApproxStructure.build((s, str, arr) => {
         return s.element('div', {
           classes: [ arr.has('tox-tinymce'), arr.has('tox-tinymce-inline') ],
-          styles: {
-            'max-width': str.is(`${maxWidth}px`),
-          },
           children: [
             s.element('div', {
               classes: [ arr.has('tox-editor-container') ],
               children: [
                 s.element('div', {
                   classes: [ arr.has('tox-editor-header') ],
+                  styles: {
+                    'max-width': str.is(`${maxWidth}px`),
+                  },
                   children: [
                     s.element('div', {
-                      classes: [ arr.has('tox-toolbar') ],
+                      classes: [ arr.has('tox-toolbar-overlord') ],
                       attrs: { role: str.is('group') }
                     }),
                     s.element('div', {
@@ -46,32 +49,69 @@ UnitTest.asynctest('Inline Editor (Silver) width test', (success, failure) => {
     )
   ]));
 
-  const sAssetWidth = (uiContainer: Element, maxWidth: number) => Chain.asStep(uiContainer, [
-    UiFinder.cFindIn('.tox-toolbar'),
+  const sAssetWidth = (uiContainer: Element, maxWidth: number, minWidth: number = 0) => Chain.asStep(uiContainer, [
+    UiFinder.cFindIn('.tox-toolbar-overlord'),
     Chain.op((toolbar) => {
       const widthString = Css.get(toolbar, 'width') || '0px';
       const width = parseInt(widthString.replace('px', ''), 10);
       Assertions.assertEq(`Toolbar with should be less than ${maxWidth}px - ${width}<=${maxWidth}`, true, width <= maxWidth);
+      Assertions.assertEq(`Toolbar with should be greater than ${minWidth}px - ${width}>=${minWidth}`, true, width >= minWidth);
     })
   ]);
 
-  const width = 400;
-  TinyLoader.setup((editor, onSuccess, onFailure) => {
-      const uiContainer = Element.fromDom(editor.getContainer());
-      const tinyApis = TinyApis(editor);
+  const sTestRender = (label: string, settings: Record<string, any>, expectedWidth: number, additionalSteps: (editor: Editor, apis) => Step<any, any>[] = Fun.constant([])) => {
+    return Step.label(label, Step.raw((_, done, die, logs) => {
+      TinyLoader.setup((editor, onSuccess, onFailure) => {
+          const uiContainer = Element.fromDom(editor.getContainer());
+          const tinyApis = TinyApis(editor);
 
-      Pipeline.async({}, [
-        tinyApis.sFocus,
-        sStructureTest(editor, uiContainer, width),
-        sAssetWidth(uiContainer, width)
-      ], onSuccess, onFailure);
-    },
-    {
-      theme: 'silver',
-      menubar: false,
-      inline: true,
-      width,
-      base_url: '/project/tinymce/js/tinymce'
-    }, success, failure
-  );
+          Pipeline.async({}, [
+            Step.sync(() => Scroll.to(0, 0)),
+            tinyApis.sFocus,
+            sStructureTest(editor, uiContainer, expectedWidth),
+            sAssetWidth(uiContainer, expectedWidth, expectedWidth - 100),
+            tinyApis.sSetContent(Arr.range(100, () => '<p></p>').join('')),
+            Step.sync(() => Scroll.to(0, 500)),
+            UiFinder.sWaitForVisible('Wait to be docked', Body.body(), '.tox-tinymce--toolbar-sticky-on .tox-editor-header'),
+            sAssetWidth(uiContainer, expectedWidth, expectedWidth - 100),
+            ...additionalSteps(editor, tinyApis)
+          ], onSuccess, onFailure, logs);
+        },
+        {
+          theme: 'silver',
+          menubar: false,
+          inline: true,
+          base_url: '/project/tinymce/js/tinymce',
+          toolbar_drawer: 'floating',
+          ...settings
+        }, done, die
+      );
+    }));
+  };
+
+  Pipeline.async({}, [
+    sTestRender('Check max-width is 400px when set via init', { width: 400 }, 400),
+    sTestRender('Check max-width is 400px when set via element', {
+      setup: (ed: Editor) => {
+        Css.set(Element.fromDom(ed.getElement()), 'width', '400px');
+      }
+    }, 400),
+    sTestRender('Check max-width is constrained to the body width when no width set', {
+      setup: (ed: Editor) => {
+        ed.on('PreInit', () => {
+          Css.set(Body.body(), 'width', '400px');
+        });
+        ed.on('Remove', () => {
+          Css.remove(Body.body(), 'width');
+        });
+      }
+    }, 400),
+    sTestRender('Check width when expanding sliding toolbar while docked', {
+      toolbar_drawer: 'sliding',
+      width: 400
+    }, 400, (editor) => [
+      sOpenMore(ToolbarDrawer.sliding),
+      sAssetWidth(Element.fromDom(editor.getContainer()), 400, 300)
+    ]),
+  ], success, failure);
 });